extract color functions to helpers/colors
